[feature] CurrentRateTile: renovate tile layout and rename functions
